Make atransport be a real class.
Using non-POD types in atransport means we'll need to start treating
it as a real class (specifically with regards to new/delete rather
than malloc/free).

I've also cleaned up the home grown linked lists for transport_list
and pending_list to just be std::lists. We might want to refactor that
again to be an std::unordered_map keyed on serial, since that seems to
be a common way to search it.

Cherry-picked from
Git-Repo: https://android.googlesource.com/platform/system/core/+/refs/heads/nougat-release
Git-Commit: c7915a3470292349017f94ca066ed515babfcc23

Change-Id: I7f5e23cdc47944a9278099723ca029585fe52105