system: core: usb: Add support for F005 USB composition on cinder
Add support for serial console over usb on cinder.

0xF005 - Standalone CDC/ACM

F005 composition script was added for this purpose. A new ttyGS node
will be created in /dev after acm.0 is created in configfs. Agetty
service will be running in background which opens the ttyGS0 port
and spawns a new console over it.

Change-Id: I5af1baf528f305677d36045839498589bce57e69
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>