adbd: update send_file function to truncate files that failed to unlink
Detect the case where file still exists even though it should have been
unlinked. Unlink can fail in some cases such as a bind mount to a file
on a read-only filesystem. This can lead to unwanted data in the
destination file if the source file is smaller than the existing file.
In such case, open with O_TRUNC option so that file starts with zero
size as expected.

Change-Id: Ib930eddd503c911733f4e678178d758d8225fd65