system: core: adb: fix adb atransport null pointer function calls.
When the close() and kick() function pointers of class
atransport are called, they are not determined whether
they are null,but when using network socket adb, the
close() and kick() function pointers are null pointers,
which will cause the null pointer function to be called
when the network adb is disconnected.

Assign default functions to the close() and kick()
pointers in the class atransport constructor.

Change-Id: Iccc201f74c84b57af248a4a0ef79e4507b223691
Signed-off-by: Xinqi Zhang <quic_xinqzhan@quicinc.com>